refactored the code

+++ crud/note.py
@@ -1,44 +1,44 @@
 
-from typing import Annotated
-
 from psycopg import Connection
 
-from notes_api.schemas.notes import Notes
+from notes_api.schemas.notes import Notes, NotesUpdate
 
-async def crud_get_one(id: int, conn: Connection):
-    sql = """SELECT * FROM notes WHERE note_id = %s"""
+async def crud_get_one(email: str, conn: Connection):
+    #sql = """SELECT * FROM notes WHERE note_id = %s"""
+    sql = """SELECT  FROM notes WHERE email = %s FETCH NEXT 1 ROWS ONLY"""
     with conn.cursor() as cur:
-        cur.execute(sql, (id,))
+        cur.execute(sql, (email,))
         val = cur.fetchone()
         return val
     
-async def crud_retrieve(conn: Connection):
-    sql = """SELECT * FROM notes;"""
+async def crud_retrieve(user_id: int, conn: Connection):
+    sql = """SELECT * FROM notes WHERE user_id = %s;"""
     with conn.cursor() as cur:
-        cur.execute(sql)
-        return cur.fetchall()
-        
-async def crud_create(note: Notes, conn: Connection):
-    sql = """INSERT INTO notes (title, content) VALUES (%s, %s) RETURNING *"""
+        cur.execute(sql, (user_id,))
+        res = cur.fetchall()
+        return res
+    
+async def crud_create(user_id: int, note: Notes, conn: Connection):
+    sql = """INSERT INTO notes (user_id, title, content) VALUES (%s, %s, %s) RETURNING *"""
     with conn.cursor() as cur:
-        #if not note.title: 
-         #   note.title = note.title[:58]
-        cur.execute(sql, (note.title, note.content,))
+        if not note.title: 
+           note.title = note.content[:57]
+           print(note.title)
+        cur.execute(sql, (user_id, note.title, note.content,))
         
-async def crud_update(note_id: int, note: Notes, conn: Connection):
-    sql = """UPDATE notes SET content = %s WHERE note_id = %s RETURNING *;""" 
+async def crud_update(user_id: int, note_id, note: NotesUpdate, conn: Connection):
+    sql = """UPDATE notes SET content = %s WHERE user_id = %s AND note_id = %s RETURNING *;""" 
     with conn.cursor() as cur:
         if note.title:
-            sql = """UPDATE notes SET title = %s, content = %s WHERE note_id = %s RETURNING *;"""       
-            cur.execute(sql, (note.title, note.content, note_id,))
+            sql = """UPDATE notes SET title = %s, content = %s WHERE user_id = %s AND note_id = %s RETURNING *;"""       
+            cur.execute(sql, (note.title, note.content, user_id, note_id,))
         else:
-            cur.execute(sql, (note.content, note_id,))
+            cur.execute(sql, (note.content, user_id, note_id,))
         return cur.fetchone()
        
-    
-async def crud_delete(note_id: int, conn: Connection):
-    sql = """DELETE FROM notes WHERE note_id = %s RETURNING *;"""
+async def crud_delete(user_id: int, note_id: int, conn: Connection):
+    sql = """DELETE FROM notes WHERE user_id = %s AND note_id = %s RETURNING *;"""
     with conn.cursor() as cur:
-        cur.execute(sql, (note_id,))
+        cur.execute(sql, (user_id, note_id,))
         data = cur.fetchone()
     return data
